Adjusted image.controller
